+++ src/scripts/mount.ts
@@ -55,22 +55,30 @@ namespace Scripts {
             const shrinkedMountsGraphic = '0x211F|0x2121|0x2124|0x20F6|0x2120|0x2135|0x2136|0x2137|0x20DD';
             const shrinkedMounts = Orion.FindType(shrinkedMountsGraphic, '0xFFFF');
             const mountsGraphic = '0x00DF|0x00DC|0x00DA|0x00E2|0x00CC|0x00E4|0x00D2|0x00DB|0x00C8';
-            for (const m of shrinkedMounts) {
-                Orion.ClearJournal();
-                Orion.UseObject(m);
-                Orion.Wait(responseDelay);
-                if (Orion.InJournal('You cannot unshrink creature here or now')) {
-                    Scripts.Utils.log('You cannot unshrink creature here or now', ColorEnum.red);
-                    throw 'e';
-                }
-                const groundPets = Orion.FindType(mountsGraphic, '0xFFFF', 'ground', 'live', 5);
-                for (const pet of groundPets) {
-                    const petObject = Orion.FindObject(pet);
-                    if (petObject.CanChangeName()) {
-                        Orion.AddObject('myMount', pet);
-                        Orion.UseObject(pet);
-                        break;
-                    }
+
+            if (!shrinkedMounts.length) {
+                return;
+            }
+
+            Orion.WarMode(true);
+            Orion.ClearJournal();
+            Orion.UseObject(shrinkedMounts[0]);
+            Orion.Wait(responseDelay);
+            if (Orion.InJournal('You cannot unshrink creature here or now')) {
+                Orion.ClearJournal('You cannot unshrink creature');
+                Scripts.Utils.log('You cannot unshrink creature here or now', ColorEnum.red);
+                throw 'e';
+            }
+
+            Scripts.Klamak.klamakCooldown();
+
+            const groundPets = Orion.FindType(mountsGraphic, '0xFFFF', 'ground', 'live', 5);
+            for (const pet of groundPets) {
+                const petObject = Orion.FindObject(pet);
+                if (petObject.CanChangeName()) {
+                    Orion.AddObject('myMount', pet);
+                    Orion.UseObject(pet);
+                    break;
                 }
             }
         }